advanced_llm_apps/chat-with-tarots/requirements.txt
# Core dependencies
streamlit
pandas
langchain
langchain-core
langchain-community
dashscope
python-dotenv